Split index_derivs.cxx into multiple files
Should help with certain compilers, may improve compile times
